week_2/lab2/requirements.txt
pydriller>=2.8
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
transformers>=4.30.0
torch>=2.0.0
requests>=2.31.0
gitpython>=3.1.0
scikit-learn>=1.3.0
